# Speed up Pattern Lab compile times by 3x; speed up Travis