--- a/docs-site/.patches/faster-compile-times--add-twig-cache-support--pattern-loader.patch
+++ b/docs-site/.patches/faster-compile-times--add-twig-cache-support--pattern-loader.patch
@@ -1,0 +1,31 @@
+diff --git a/src/PatternLab/PatternEngine/Twig/Loaders/PatternLoader.php b/src/PatternLab/PatternEngine/Twig/Loaders/PatternLoader.php
+index 4fa55da73..e0de95c28 100644
+--- a/src/PatternLab/PatternEngine/Twig/Loaders/PatternLoader.php
++++ b/src/PatternLab/PatternEngine/Twig/Loaders/PatternLoader.php
+@@ -30,6 +30,7 @@ class PatternLoader extends Loader {
+ 		// set-up default vars
+ 		$twigDebug      = Config::getOption("twigDebug");
+ 		$twigAutoescape = Config::getOption("twigAutoescape");
++		$twigCache = Config::getOption("twigCache");
+ 		
+ 		// go through various places where things can exist
+ 		$filesystemLoaderPaths = array();
+@@ -88,7 +89,12 @@ class PatternLoader extends Loader {
+ 		
+ 		// set-up Twig
+ 		$twigLoader = new \Twig_Loader_Chain($loaders);
+-		$instance   = new \Twig_Environment($twigLoader, array("debug" => $twigDebug, "autoescape" => $twigAutoescape));
++		$instance   = new \Twig_Environment($twigLoader, array(
++			"debug" => false,
++			"autoescape" => $twigAutoescape,
++			"cache" => $twigCache,
++			"auto_reload" => true,
++		));
+ 		
+ 		// customize Twig
+ 		TwigUtil::setInstance($instance);
+@@ -133,3 +139,4 @@ class PatternLoader extends Loader {
+ 	}
+ 	
+ }
++