--- a/docs-site/.patches/faster-compile-times--docs--composer.json.patch
+++ b/docs-site/.patches/faster-compile-times--docs--composer.json.patch
@@ -1,0 +1,14 @@
+diff --git a/composer.json b/composer.json
+index da40fa12e..b5dea1e43 100644
+--- a/core/composer.json
++++ b/core/composer.json
+@@ -49,6 +49,8 @@
+     "symfony/filesystem": "^3.0",
+     "symfony/finder": "^3.0",
+     "symfony/process": "^3.1",
+-    "symfony/yaml": "^3.0"
++    "symfony/yaml": "^3.0",
++    "cebe/markdown": "^1.2",
++    "hyn/frontmatter": "^1.1"
+   }
+ }