--- a/docs-site/.patches/faster-compile-times--docs--documentation.patch
+++ b/docs-site/.patches/faster-compile-times--docs--documentation.patch
@@ -1,0 +1,95 @@
+diff --git a/src/PatternLab/Parsers/Documentation.php b/src/PatternLab/Parsers/Documentation.php
+index f28606210..b60048227 100644
+--- a/src/PatternLab/Parsers/Documentation.php
++++ b/src/PatternLab/Parsers/Documentation.php
+@@ -12,26 +12,15 @@
+ 
+ namespace PatternLab\Parsers;
+ 
+-use \Michelf\MarkdownExtra;
+ use \PatternLab\Timer;
+ use \Symfony\Component\Yaml\Exception\ParseException;
+ use \Symfony\Component\Yaml\Yaml;
+ 
+ class Documentation {
+-	
++	private static $parser;
++
+ 	protected static $lineEndingsSet = false;
+ 	
+-	/**
+-	* Convert markdown data into proper mark-up
+-	* @param  {String}       the text to be converted
+-	*
+-	* @return {String}       the converted mark-up
+-	*/
+-	public static function convertMarkdown($text) {
+-		$markdown = MarkdownExtra::defaultTransform($text);
+-		return $markdown;
+-	}
+-	
+ 	/**
+ 	* Parse YAML data into an array
+ 	* @param  {String}       the text to be parsed
+@@ -84,46 +73,17 @@ class Documentation {
+ 	* @return {Array}        array containing both the YAML and converted markdown
+ 	*/
+ 	public static function parse($text) {
+-		
+-		self::setLineEndings();
+-		
+-		// set-up defaults
+-		$yaml     = array();
+-		$markdown = "";
+-		
+-		// read in the content
+-		// based on: https://github.com/mnapoli/FrontYAML/blob/master/src/Parser.php
+-		$lines = explode(PHP_EOL, $text);
+-		
+-		if (count($lines) <= 1) {
+-			$markdown = self::convertMarkdown($text);
+-			return array($yaml,$markdown);
+-		}
+-		
+-		if (rtrim($lines[0]) !== '---') {
+-			$markdown = self::convertMarkdown($text);
+-			return array($yaml,$markdown);
+-		}
+-		
+-		$head = array();
+-		unset($lines[0]);
+-		$i = 1;
+-		foreach ($lines as $line) {
+-			if ($line === '---') {
+-				break;
+-			}
+-			$head[] = $line;
+-			$i++;
++
++		if (!isset(self::$parser)) {
++			self::$parser = new \Hyn\Frontmatter\Parser(new \cebe\markdown\Markdown);
++			self::$parser->setFrontmatter(\Hyn\Frontmatter\Frontmatters\YamlFrontmatter::class);
+ 		}
+-		
+-		$head = implode(PHP_EOL, $head);
+-		$body = implode(PHP_EOL, array_slice($lines, $i));
+-		
+-		$yaml     = self::convertYAML($head);
+-		$markdown = self::convertMarkdown($body);
+-		
++
++		$markdown = "";
++		$content = self::$parser->parse($text);
++		$markdown = $content['html'];
++		$yaml     = $content['meta'];
+ 		return array($yaml,$markdown);
+-		
+ 	}
+ 	
+ 	/**
+@@ -135,5 +95,4 @@ class Documentation {
+ 			self::$lineEndingsSet = true;
+ 		}
+ 	}
+-	
+ }