--- a/docs-site/.patches/faster-compile-times--ignore-more-files.patch
+++ b/docs-site/.patches/faster-compile-times--ignore-more-files.patch
@@ -1,0 +1,75 @@
+diff --git a/src/PatternLab/PatternData.php b/src/PatternLab/PatternData.php
+index 5eab40386..2c2df1d9c 100644
+--- a/src/PatternLab/PatternData.php
++++ b/src/PatternLab/PatternData.php
+@@ -127,37 +127,31 @@ class PatternData {
+ 	 *  @TODO: figure out how to store this PL's main config
+ 	 */
+ 
+-		$finder = new Finder();
+-		$patternObjects = $finder
+-		                  ->in($patternSourceDir)
+-											->followLinks()
+-											->ignoreDotFiles(true)
+-											->notName('package.json')
+-											->notName('package-lock.json')
+-											->notName('README.md')
+-											->notName('*.js')
+-											->notName('*.map')
+-											->notName('*.lock')
+-											->notName('*.scss')
+-											->notName('*.css')
+-											->notName('_*')
+-											->exclude(array('node_modules', '.git', 'tests'));
+- 		$patternObjects = iterator_to_array($patternObjects);
+-
+-
+-		// Only care about README.md files that contain front-matter
+-		$finderReadmeDocs = new Finder();
+-		$patternReadmeDocs = $finderReadmeDocs
+-											->in($patternSourceDir)
+-											->name('README.md')
+-											->contains('/^(---)/')
+-											->notName('_*')
+-											->exclude(array('node_modules', '.git', 'tests'));
+-		$patternReadmeDocs = iterator_to_array($patternReadmeDocs);
+-
+-		// Merge together the default set of patterns found + the README.md-filtered patterns.
+-		//$patternObjects = array_merge_recursive($patternObjects, $patternReadmeDocs);
+-
++	$finder = new Finder();
++	$patternObjects = $finder
++										->in($patternSourceDir)
++										->followLinks()
++										->ignoreDotFiles(true)
++										->notName('package.json')
++										->notName('package-lock.json')
++										->notName('README.md')
++										->notName('CHANGELOG.md')
++										->notName('TESTING.md')
++										->notName('*.js')
++										->notName('*.snap')
++										->notName('*.yml')
++										->notName('*.yaml')
++										->notName('*.php')
++										->notName('*.png')
++										->notName('*.svg')
++										->notName('*.jpg')
++										->notName('*.map')
++										->notName('*.lock')
++										->notName('*.scss')
++										->notName('*.css')
++										->notName('_*.*')
++										->exclude(array('node_modules', 'vendor', '.git', '__tests__', 'tests'));
++	 $patternObjects = iterator_to_array($patternObjects);
+ 
+ 		// sort the returned objects
+ 		ksort($patternObjects);
+@@ -183,7 +177,6 @@ class PatternData {
+ 					$rule->run($depth, $ext, $path, $pathName, $name);
+ 				}
+ 			}
+-
+ 		}
+ 
+ 		// dispatch that the data is loaded