--- a/docs-site/composer.json
+++ b/docs-site/composer.json
@@ -33,7 +33,9 @@
     "pattern-lab/core": "dev-develop as 2.9.0",
     "pattern-lab/patternengine-twig": "^2.2.2",
     "cweagans/composer-patches": "^1.6.4",
-    "evanlovely/plugin-twig-namespaces": "^1.1.1"
+    "evanlovely/plugin-twig-namespaces": "^1.1.1",
+    "cebe/markdown": "^1.2",
+    "hyn/frontmatter": "^1.1"
   },
   "autoload": {
     "psr-0": {