--- a/docs-site/config/config.yml
+++ b/docs-site/config/config.yml
@@ -5,7 +5,7 @@ ie:
     - less
     - scss
     - js
-    - json #so json files don't get copied over to the www folder
+    - json
 id:
     - scss
     - .svn
@@ -18,9 +18,6 @@ patternStates:
     - inprogress
     - inreview
     - complete
-# styleGuideExcludes:
-    # - 04-pages
-    # - viewall-pages-all
 cacheBusterOn: 'true'
 publicDir: ../www/pattern-lab/
 sourceDir: src/pages/pattern-lab/
@@ -50,11 +47,11 @@ lineageMatch: '{%([ ]+)?(?:include|extends|embed)( |\()[&quot;\&#039;]([\/.@A-Za
 lineageMatchKey: 3
 patternExtension: twig
 twigExtensions:
-    - '\BasaltInc\TwigTools\TwigExtensions\BasaltFakerExtension'
-    - '\Bolt\TwigExtensions\BoltCore'
-    - '\Bolt\TwigExtensions\BoltCoreCompat'
-    - '\Bolt\TwigExtensions\BoltExtras'
-twigDebug: true
+    - \BasaltInc\TwigTools\TwigExtensions\BasaltFakerExtension
+    - \Bolt\TwigExtensions\BoltCore
+    - \Bolt\TwigExtensions\BoltCoreCompat
+    - \Bolt\TwigExtensions\BoltExtras
+twigCache: cache/twig
 twigAutoescape: false
 twigDefaultDateFormat: ''
 twigDefaultIntervalFormat: ''
@@ -70,17 +67,15 @@ extensions:
     docs: .docs.md
     sg: .md
 plugins:
-  twigNamespaces:
-    enabled: true
-    readFromFile: ../www/build/data/twig-namespaces.bolt.json
-# @todo: remove once we move to Pattern Lab Node + move this config into our patternlab.config.json
+    twigNamespaces:
+        enabled: true
+        readFromFile: ../www/build/data/twig-namespaces.bolt.json
 theme:
-  logo:
-    srcDark: '/images/bolt-logo--dark.svg'
-    srcLight: '/images/bolt-logo--light.svg'
-    url: '/pattern-lab/?p=components-overview'
-    altText: 'Logo of the Bolt Design System'
-    text: ''
-  color: 'dark'
-  layout: 'vertical'
-  # "density" : "compact" | "cozy" | "comfortable",
+    logo:
+        srcDark: /images/bolt-logo--dark.svg
+        srcLight: /images/bolt-logo--light.svg
+        url: '/pattern-lab/?p=components-overview'
+        altText: 'Logo of the Bolt Design System'
+        text: ''
+    color: dark
+    layout: vertical