--- a/docs-site/config/config.yml
+++ b/docs-site/config/config.yml
@@ -79,3 +79,4 @@ theme:
         text: ''
     color: dark
     layout: vertical
+twigDebug: false