--- a/packages/twig-integration/twig-renderer/SetupTwigRenderer.php
+++ b/packages/twig-integration/twig-renderer/SetupTwigRenderer.php
@@ -5,8 +5,8 @@ require_once 'vendor/autoload.php';
 function addBoltCoreExtensions(\Twig\Environment &$env, $config) {
   $env->addExtension(new \Bolt\TwigExtensions\BoltCore());
   $env->addExtension(new \Bolt\TwigExtensions\BoltCoreCompat());
-  $env->addExtension(new \Twig_Extension_Debug());
-  $env->addExtension(new \BasaltInc\TwigTools\TwigExtensions\BasaltFakerExtension());
+  // $env->addExtension(new \Twig_Extension_Debug());
+  // $env->addExtension(new \BasaltInc\TwigTools\TwigExtensions\BasaltFakerExtension());
 
   $env->addFunction(new \Twig_SimpleFunction('customTwigFunctionThatSaysWorld', function () {
     return 'Custom World';