## Jira
http://vjira2:8080/browse/BDS-2076

## Summary
_Significantly_ speeds up Pattern Lab's compile times through a number of performance-related patches.

### On Travis:
- **Before:**
    - Pattern Lab Data compiled in ~`12.4s`
    - Pattern Lab compiled in ~`1m 57.4s`
- **After:**
    - Pattern Lab Data compiles in ~`605ms` (or ~`11.8s` faster)
    - Pattern Lab compiles in ~`46s` (or about `71.4s` faster)

### Locally (running the `php core/console --generate` command)
- **Before:** 
    - Pattern Lab compiles in ~`33.5s`
- **After**:
    - Pattern Lab compiles in ~`12s` (no cache)  --> about 2x faster!
    - Pattern Lab compiles/recompiles in ~`8.8s` (with cache)  --> more than 3.5x faster!

<hr>

## Details
- Patches Pattern Lab PHP to add support for Twig's built-in caching + enables `auto_reload` by default
- Disables debug mode in Pattern Lab (ie. no more `{{ dump() }}`) in order to gain significant performance improvements when rendering Twig
- Updates the original Pattern Lab code handling markdown and front-matter parsing with a much faster off the shelf library.
- Updates our Faker.php Twig extension to support PHP 7.4 without breaking PHP 7.3 compatibility
- Blacklists additional files for Pattern Lab to ignore parsing

## How to test
- Review changes
- Confirm Pattern Lab compile times have significantly improved when running locally